[#6] Deploy hokuto_flask as a serverless app
With this commit the Flask app hokuto_flask can be deployed on AWS API
Gateway and AWS Lambda with Zappa.

Since the Heroku Dyno used to host this app is no longer required, this
commit also removes the `Procfile`.

IMPORTANT: even if the Heroku Dyno is no longer used, the PostgreSQL
database installed as add-on for that Dyno is still the only production
database available, so don't drop it yet!