[#26] Remove psycopg2-binary (use a lambda layer)
This commit removes psycopg2-binary from the deployment package.

Note that at the moment Zappa does not support lambda layers (though it
probably will in the near future), so for the time being we need to
manually declare the lambda layer our lambda requires. This can be done
on the AWS Lambda website, and maybe also from the aws CLI.

See:

- https://github.com/Miserlou/Zappa/issues/1716
- https://github.com/Miserlou/Zappa/pull/1752